Moved code into dev.py
Moved runSimulations() to dev.py from main.py
